added WizardHelper class to be shared with other Wizards (#6033)
* added WizardHelper class to be shared with other Wizards

* Update src/openms_gui/include/OpenMS/VISUAL/DIALOGS/sources.cmake
