+++ src/openms_gui/include/OpenMS/VISUAL/DIALOGS/sources.cmake
@@ -28,6 +28,7 @@ TOPPViewOpenDialog.h
 TOPPViewPrefDialog.h
 TheoreticalSpectrumGenerationDialog.h
 ToolsDialog.h
+WizardHelper.h
 )
 
 ### add path to the filenames
