+++ src/openms_gui/source/VISUAL/DIALOGS/SwathTabWidget.cpp
@@ -34,14 +34,13 @@
 
 #include <OpenMS/VISUAL/DIALOGS/SwathTabWidget.h>
 #include <ui_SwathTabWidget.h>
+#include <OpenMS/VISUAL/DIALOGS/WizardHelper.h>
 
 #include <OpenMS/CONCEPT/LogStream.h>
 #include <OpenMS/FORMAT/FileHandler.h>
 #include <OpenMS/FORMAT/ParamXMLFile.h>
 #include <OpenMS/SYSTEM/File.h>
 #include <OpenMS/VISUAL/DIALOGS/PythonModuleRequirement.h>
-#include <OpenMS/VISUAL/DIALOGS/TOPPASInputFilesDialog.h>
-#include <OpenMS/VISUAL/MISC/GUIHelpers.h>
 
 #include <QtCore/QDateTime>
 #include <QtCore/QDir>
@@ -58,20 +57,7 @@ namespace OpenMS
 {
   namespace Internal
   {
-
-    SwathGUILock::SwathGUILock(SwathTabWidget* stw)
-      : 
-      stw_(stw),
-      old_(stw->currentWidget()),
-      glock_(stw)
-    {
-      stw->setCurrentWidget(stw->ui->tab_log);
-    }
-
-    SwathGUILock::~SwathGUILock()
-    {
-      stw_->setCurrentWidget(old_);
-    }
+    template class WizardGUILock<SwathTabWidget>;
 
     String getOSWExe()
     {
@@ -179,8 +165,8 @@ namespace OpenMS
     void SwathTabWidget::on_run_swath_clicked()
     {
       if (!checkOSWInputReady_()) return;
-      
-      SwathGUILock lock(this); // forbid user interaction
+
+      WizardGUILock lock(this); // forbid user interaction
 
       updateSwathParamFromWidgets_();
       Param tmp_param;
@@ -385,63 +371,6 @@ namespace OpenMS
       ui->input_swath_windows->setCWD(new_cwd);
     }
 
-    /// custom arguments to allow for looping calls
-    struct Args
-    {
-      QStringList loop_arg; ///< list of arguments to insert; one for every loop
-      size_t insert_pos;       ///< where to insert in the target argument list (index is 0-based)
-    };
-    
-    typedef std::vector<Args> ArgLoop;
-
-    /// Allows running an executable with arguments
-    /// Multiple execution in a loop is supported by the ArgLoop argument
-    /// e.g. running 'ls -la .' and 'ls -la ..'
-    /// uses Command("ls", QStringList() << "-la" << "%1", ArgLoop{ Args {QStringList() << "." << "..", 1 } })
-    /// All lists in loop[i].loop_arg should have the same size (i.e. same number of loops)
-    struct Command
-    {
-      String exe;
-      QStringList args;
-      ArgLoop loop;
-
-      Command(const String& e, const QStringList& a, const ArgLoop& l) :
-        exe(e),
-        args(a),
-        loop(l) {}
-
-      /// how many loops can we make according to the ArgLoop provided?
-      /// if ArgLoop is empty, we just do a single invokation
-      size_t getLoopCount() const
-      {
-        if (loop.empty()) return 1;
-        size_t common_size = loop[0].loop_arg.size();
-        for (const auto& l : loop)
-        {
-          if (l.loop_arg.size() != (int)common_size) throw Exception::Precondition(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Internal error. Not all loop arguments support the same number of loops!");
-          if ((int)l.insert_pos >= args.size()) throw Exception::Precondition(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Internal error. Loop argument wants to insert after end of template arguments!");
-        }
-        return common_size;
-      }
-      /// for a given loop, return the substituted arguments
-      /// @p loop_number of 0 is always valid, i.e. no loop args, just use the unmodified args provided
-      QStringList getArgs(const int loop_number) const
-      {
-        if (loop_number >= (int)getLoopCount())
-        {
-          throw Exception::Precondition(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Internal error. The loop number you requested is too high!");
-        }
-        if (loop.empty()) return args; // no looping available
-
-        QStringList arg_l = args;
-        for (const auto& largs : loop) // replace all args for the current round
-        {
-          arg_l[largs.insert_pos] = args[largs.insert_pos].arg(largs.loop_arg[loop_number]);
-        }
-        return arg_l;
-      }
-    };
-
     bool SwathTabWidget::findPythonScript_(const String& path_to_python_exe, String& script_name)
     {
       String path = File::path(path_to_python_exe);
@@ -476,7 +405,7 @@ namespace OpenMS
         QMessageBox::warning(this, "Error", "Could not find all requirements for 'pyprophet & tric' (see 'Config' tab). Install modules via 'pip install <modulename>' and make sure it's available in $PATH");
         return;
       }
-      SwathGUILock lock(this); // forbid user interaction
+      WizardGUILock lock(this); // forbid user interaction
 
       auto inputs = getPyProphetInputFiles();
       if (inputs.empty())
